fix and refactor css, fonts, card component
