+++ src/components/PizzaCard/PizzaCard.jsx
@@ -1,6 +1,6 @@
-import pizza from "../../assets/img/pizza.png";
+import pizza from '../../assets/img/pizza.png';
 
-import cl from "./PizzaCard.module.css";
+import cl from './PizzaCard.module.scss';
 
 const PizzaCard = () => {
   return (
@@ -10,14 +10,21 @@ const PizzaCard = () => {
         <h3 className={cl.title}>Чизбургер-пицца</h3>
         <div className={cl.variants}>
           <ul className={cl.list}>
-            <li className={cl.item + " " + cl.active}>тонкое</li>
+            <li className={cl.item + ' ' + cl.active}>тонкое</li>
             <li className={cl.item}>традиционное</li>
           </ul>
           <ul className={cl.list}>
             <li className={cl.item}>26 см.</li>
             <li className={cl.item}>30 см.</li>
-            <li className={cl.item + " " + cl.active}>40 см.</li>
+            <li className={cl.item + ' ' + cl.active}>40 см.</li>
           </ul>
+        </div>
+        <div className={cl.footer}>
+          <div className={cl.price}>от 395 ₽</div>
+          <button className={cl.btn}>
+            + добавить
+            <span>2</span>
+          </button>
         </div>
       </div>
     </div>
