--- src/components/PizzaCard/PizzaCard.module.css
@@ -1,46 +1,0 @@
-.card {
-  text-align: center;
-}
-
-.img {
-  margin-inline: auto;
-}
-
-.title {
-  font-weight: var(--fw-eb);
-  font-size: var(--fz-500);
-}
-
-.info > * + * {
-  margin-top: 1.5rem;
-}
-
-.variants {
-  padding: .5rem;
-  background-color: var(--clr-neutral-150);
-  border-radius: .625rem;
-}
-
-.list {
-  display: flex;
-}
-
-.list > * {
-  flex: 1;
-}
-
-.list + .list {
-  margin-top: .5rem;
-}
-
-.item {
-  border-radius: .3125rem;
-  padding: .5rem;
-  cursor: pointer;
-  font-size: var(--fz-300);
-  font-weight: var(--fw-bold);
-}
-
-.active {
-  background-color: #fff;
-}
