+++ src/pages/Home/Home.jsx
@@ -1,7 +1,7 @@
-import Filters from "../../components/Filters/Filters";
-import Pizzas from "../../components/Pizzas/Pizzas";
+import Filters from '../../components/Filters/Filters';
+import Pizzas from '../../components/Pizzas/Pizzas';
 
-import cl from "./Home.module.css";
+import cl from './Home.module.scss';
 
 const Home = () => {
   return (
