--- src/pages/Home/Home.module.css
@@ -1,7 +1,0 @@
-.main {
-  padding: 2.5rem;
-}
-
-.main > * + * {
-  margin-top: 2rem;
-}
